+++ profiling/pipeline_generator.py
@@ -1,3 +1,5 @@
+from typing import Optional
+
 from sklearn.pipeline import Pipeline
 
 from tubular.capping import CappingTransformer, OutOfRangeNullTransformer
@@ -89,7 +91,7 @@ class TubularPipelineGenerator:
 
     def generate_pipeline(
         self,
-        transformers_to_include: list = None,
+        transformers_to_include: Optional[list] = None,
         verbose: bool = False,
     ) -> Pipeline:
         if not transformers_to_include: